Un poco mas de API's en Article

+++ Proyecto24AM/Models/Entities/Article.cs
@@ -7,6 +7,6 @@ namespace Proyecto24AM.Models.Entities
         [Key]public int PKArticle { get; set; }
         [Required]public string Name { get; set; }
         [Required] public string Description { get; set; }
-        [Required] public decimal Price { get; set; }
+        [Required] public double Price { get; set; }
     }
 }
